+++ internal/handlers/taskhandlers/task_done.go
@@ -52,11 +52,8 @@ func TaskDone(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	byteResp, err := json.Marshal(map[string]string{})
-	if err != nil {
+	if err = json.NewEncoder(w).Encode(map[string]string{}); err != nil {
 		writeErrorInJson(w, err)
 		return
 	}
-
-	w.Write(byteResp)
 }
