+++ internal/nextdate/nextdate.go
@@ -2,6 +2,9 @@ package nextdate
 
 import (
 	"time"
+
+	count_nextdate "github.com/kasariks/project_for_graduating/internal/nextdate/count"
+	validate_nextdate "github.com/kasariks/project_for_graduating/internal/nextdate/validate"
 )
 
 const DateFormat = "20060102"
@@ -11,11 +14,12 @@ func NextDate(now time.Time, dstart string, repeat string) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	if err := validateRepeatParam(repeat); err != nil {
+
+	if err := validate_nextdate.ValidateRepeatParam(repeat); err != nil {
 		return "", err
 	}
 
-	countedTime, err := countNextDate(now, date, repeat)
+	countedTime, err := count_nextdate.CountNextDate(now, date, repeat)
 	if err != nil {
 		return "", err
 	}
