+++ internal/nextdate/validate/validateM.go
@@ -1,99 +1,10 @@
-package nextdate
+package validate_nextdate
 
 import (
 	"errors"
-	"slices"
 	"strconv"
 	"strings"
 )
-
-func validateRepeatParam(repeat string) error {
-	firstLetters := []string{"d", "y", "w", "m"}
-	splitedRepeat := splitRepeat(repeat)
-	if len(splitedRepeat) < 1 {
-		return errors.New("empty repeat")
-	}
-
-	option := splitedRepeat[0]
-
-	if !slices.Contains(firstLetters, option) {
-		return errors.New("unsupported option")
-	}
-
-	if err := validateRepeatbyOption(option, repeat); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func validateRepeatbyOption(option string, repeat string) error {
-	switch option {
-	case "d":
-		return validateRepeatD(repeat)
-	case "y":
-		return validateRepeatY(repeat)
-	case "w":
-		return validateRepeatW(repeat)
-	case "m":
-		return validateRepeatM(repeat)
-	default:
-		return errors.New("unsupported option")
-	}
-}
-
-func validateRepeatD(repeat string) error {
-	splitedRepeat := splitRepeat(repeat)
-	if len(splitedRepeat) != 2 {
-		return errors.New("unsupported format length for d option")
-	}
-
-	return validateNumbersForD(splitedRepeat[1])
-}
-
-func validateNumbersForD(number string) error {
-	n, err := strconv.Atoi(number)
-	if err != nil {
-		return errors.New("not numbers in repeat for d option")
-	}
-	if n > 400 {
-		return errors.New("too big number for d option")
-	}
-
-	return nil
-}
-
-func validateRepeatY(repeat string) error {
-	splitedRepeat := splitRepeat(repeat)
-	if len(splitedRepeat) > 1 {
-		return errors.New("unsupported format length for y option")
-	}
-
-	return nil
-}
-
-func validateRepeatW(repeat string) error {
-	splitedRepeat := splitRepeat(repeat)
-	if len(splitedRepeat) < 2 {
-		return errors.New("unsupported format length for w option")
-	}
-
-	return validateNumbersForW(splitedRepeat[1:])
-}
-
-func validateNumbersForW(numbers []string) error {
-	for _, v := range numbers {
-		n, err := strconv.Atoi(v)
-		if err != nil {
-			return errors.New("not numbers in repeat for w option")
-		}
-		if n > 7 || n < 1 {
-			return errors.New("incorrect numbers in repeat for w option")
-		}
-	}
-
-	return nil
-}
 
 func validateRepeatM(repeat string) error {
 	splitedRepeatOnlySpaces := strings.Split(repeat, " ")
